modify(Post): clap long으로 변경

+++ backend/src/main/kotlin/org/rookedsysc/mediumclone/post/Post.kt
@@ -24,7 +24,7 @@ data class Post(
     var content: String,
 
     @Column(nullable = false)
-    var clap: Int = 0,
+    var clap: Long = 0,
 
     @ManyToMany(cascade = [CascadeType.PERSIST])
     @JoinTable(
